+++ tests/unit/test_internal_deprecations.py
